backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyPDF2
python-dotenv
groq
scrapy
selenium
requests
beautifulsoup4
lxml
apscheduler


